Feature: rotten scores on imdb. Environment: Strict mode.
First working feature! :-)\n - [x] Background script added for fetching the rotten tomatoes page and getting the scores, and they are injected in the imdb page in an ugly way.\n - [ ] No full unit tests yet: no TDD. :-( But I'm learning a lot how to work in this environent, so after writing the tests, I can continue with TDD, using the acquired knowledge.\n - [x] Strict mode is enabled, for all the files. Result: no undeclared vars. Global functions have to be added to the window object, which has to be defined in the test environment.\n - [x] Eslint updated: strict mode and variable declaration is forced.